+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Legenda for the issue kind:
 
 - [#538](https://github.com/ariadne-cps/ariadne/issues/538) (N) Introduce a concurrency module designed for parallel execution of internal tasks
 - [#549](https://github.com/ariadne-cps/ariadne/issues/549) (N) Introduce a Least Recently Used cache utility for holding a limited number of homogeneous objects 
+- [#552](https://github.com/ariadne-cps/ariadne/issues/552) (C) Remove evolve/reach/reach_evolve methods from EvolverInterface, relying on orbit generation only
 - [#539](https://github.com/ariadne-cps/ariadne/issues/539) (F) A segmentation fault sometimes would be issued when terminating the executable, due to logging
 
 ### 2.2